feat: Switch to the onekey server
