+++ packages/engine/src/vaults/impl/kaspa/sdk/clientRestApi.ts
@@ -2,8 +2,8 @@ import Axios from 'axios';
 import { get } from 'lodash';
 
 import { OneKeyInternalError } from '@onekeyhq/engine/src/errors';
-
 import debugLogger from '@onekeyhq/shared/src/logger/debugLogger';
+
 import { submitTransactionFromString } from './transaction';
 
 import type { GetTransactionResponse, UTXOResponse } from './types';
